requirements.txt
# 项目依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0

# 数据处理
markdown-to-json>=2.0.0

# 可选依赖（用于高级功能）
# selenium>=4.0.0  # 用于动态网页爬取
# playwright>=1.40.0  # 另一个动态网页爬取工具

# 核心生成系统依赖
pyyaml>=6.0
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0  # GPU版本用 faiss-gpu
numpy>=1.24.0

# 可选依赖
# torch>=2.0.0  # sentence-transformers需要，如果已安装可忽略
# pypdf>=3.0.0  # PDF阅读（已在其他地方安装）
